auto-collapse tx list, refactor txlist component
